+++ examples/windows-vm/files/script-openssh.ps1
@@ -1,0 +1,3 @@
+#ps1_sysnative
+Add-WindowsCapability -Online -Name OpenSSH.Client~~~~0.0.1.0
+Add-WindowsCapability -Online -Name OpenSSH.Server~~~~0.0.1.0