+++ examples/windows-vm/settings.tf
@@ -1,0 +1,8 @@
+terraform { required_version = ">= 0.13" }
+
+locals {
+  vm_list_windows = {
+    vm1 = var.windows_vm
+    # Optionally add more VMs
+  }
+}