+++ settings.tf
@@ -13,9 +13,8 @@ locals {
   cpus       = var.cpus
   memory     = var.memory
   disk       = var.disk
-  # CloudInit ovf datasource
-  vapp_userdata = var.vapp_userdata
-  # CloudInit vmware-guestinfo datasource
-  metadata   = var.metadata
-  userdata   = var.userdata
+  # Cloud-Init datasource settingd
+  vapp_properties = var.vapp_properties
+  metadata        = var.metadata
+  userdata        = var.userdata
 }